requirements.txt
streaml
pillow